/about is /index now, moved rest pages to /legacy
